Full record for only observ fixed.

+++ src/lerobot/robots/brewie/Brewie_base.py
@@ -192,7 +192,7 @@ class BrewieBase(Robot):
     @property
     def _motors_ft(self) -> dict[str, type]:
         # Return features for all joints defined in servo_mapping
-        return {f"{joint}.pos": int for joint in self.config.servo_mapping.values()}
+        return {f"{joint}.pos": float for joint in self.config.servo_mapping.values()}
 
     @property
     def _cameras_ft(self) -> dict[str, tuple]:
@@ -322,8 +322,8 @@ class BrewieBase(Robot):
                     position = pos_data['position']
                     joint_name = self.config.servo_mapping.get(servo_id)
                     if joint_name:
-                        # Ensure position is integer and within valid range (0-1000)
-                        position = max(0, min(1000, int(position)))
+                        # Ensure position is float32 and within valid range (0-1000)
+                        position = float(max(0.0, min(1000.0, float(position))))
                         obs_dict[f"{joint_name}.pos"] = position
                         
                 with self.position_lock:
@@ -403,7 +403,7 @@ class BrewieBase(Robot):
                 servo_id = self.joint_to_id.get(joint_name)
                 if servo_id is not None:
                     # Ensure position is within valid range (0-1000)
-                    position = max(0, min(1000, int(position)))
+                    position = float(max(0.0, min(1000.0, float(position))))
                     new_positions.append({'id': servo_id, 'position': position})
             
             if new_positions:
